refactor(desktop:layout): migrate layout state from global store to per-database stores
